quick fix to remove Jaxb as dependency for terminology.
