+++ openehr-terminology/src/test/java/com/nedap/archie/terminology/OpenEHRTerminologyAccessTest.java
@@ -1,5 +1,7 @@
 package com.nedap.archie.terminology;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,6 +16,17 @@ public class OpenEHRTerminologyAccessTest {
     @Before
     public void getInstance() {
         termAccess = OpenEHRTerminologyAccess.getInstance();//should not throw an exception from parsing
+    }
+
+    @Test
+    public void writeToJson() throws Exception {
+        OpenEHRTerminologyAccess ac = OpenEHRTerminologyAccess.getInstance();
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.configure(SerializationFeature.INDENT_OUTPUT, true);
+
+        String t = mapper.writeValueAsString(ac);
+        OpenEHRTerminologyAccess parsed = mapper.readValue(t, OpenEHRTerminologyAccess.class);
+        System.out.println(t);
     }
 
     @Test
